Remove spaces from the branch names
I was about to write this exact tool, but found yours, so thank you!  As a ux/design guy the spaces before the branch name bother me and I'm not sure what their purpose is.  Given that there is a `>` cursor indicating which line you're on, and it changes the color and underlines the branch name, it seems a bit clunky.  Nitpicky, but I think it'd make for cleaner output :)